requirements.txt
absl-py==1.4.0
accelerate==0.25.0
aiofiles==22.1.0
aiosqlite==0.19.0
altgraph==0.17.4
anyio==3.7.1
arcgis==2.2.0.2
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.0
attrs==23.2.0
auto-py-to-exe==2.42.0
Babel==2.14.0
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.1.0
blinker==1.7.0
bottle==0.12.25
bottle-websocket==0.2.9
cachetools==5.3.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.2.0
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
comm==0.2.1
compel==2.0.2
contourpy==1.1.0
cryptography==41.0.7
customtkinter==5.2.1
cycler==0.11.0
darkdetect==0.8.0
dask==2023.12.1
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
diffusers==0.25.0
Eel==0.16.0
exceptiongroup==1.1.3
executing==1.2.0
fastjsonschema==2.19.1
ffmpeg==1.4
filelock==3.12.3
Flask==3.0.0
fonttools==4.42.1
fqdn==1.5.1
fsspec==2023.12.2
future==0.18.3
geomet==1.1.0
gevent==23.9.1
gevent-websocket==0.10.1
google-auth==2.23.0
google-auth-oauthlib==1.0.0
greenlet==3.0.3
grpcio==1.58.0
gssapi==1.8.3
gunicorn==21.2.0
huggingface-hub==0.20.2
idna==3.4
imageio==2.31.5
importlib-metadata==6.8.0
importlib-resources==6.0.1
ipykernel==6.28.0
ipython==8.15.0
ipython-genutils==0.2.0
ipywidgets==7.8.1
isoduration==20.11.0
itsdangerous==2.1.2
jaraco.classes==3.3.0
jedi==0.19.0
Jinja2==3.1.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
jupyter-client==6.1.12
jupyter-events==0.9.0
jupyter-server==1.24.0
jupyter-ydoc==0.2.5
jupyter_core==5.7.1
jupyter_server_fileid==0.9.1
jupyter_server_ydoc==0.8.0
jupyterlab==3.6.6
jupyterlab-widgets==1.1.7
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
keyring==24.3.0
kiwisolver==1.4.5
lazy_loader==0.3
locket==1.0.0
lxml==5.1.0
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.7.3
matplotlib-inline==0.1.6
mistune==3.0.2
more-itertools==10.2.0
mpmath==1.3.0
nbclassic==1.0.0
nbclient==0.9.0
nbconvert==7.14.0
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.1
notebook==6.5.6
notebook_shim==0.2.3
numpy==1.23.5
oauthlib==3.2.2
opencv-python==4.8.0.76
packaging==23.1
pandas==2.1.0
pandocfilters==1.5.0
parso==0.8.3
partd==1.4.1
pefile==2023.2.7
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==4.1.0
prometheus-client==0.19.0
prompt-toolkit==3.0.39
protobuf==4.21.2
psutil==5.9.5
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.16.1
pyinstaller==6.3.0
pyinstaller-hooks-contrib==2023.12
pylerc==4.0
PyOpenGL==3.1.7
pyparsing==3.1.1
pypylon==2.3.0
pyshp==2.3.1
pyspnego==0.10.2
python-certifi-win32==1.6.1
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.3.post1
PyWavelets==1.4.1
pywin32==306
pywin32-ctypes==0.2.2
pywinpty==2.0.12
PyYAML==6.0.1
pyzmq==24.0.1
referencing==0.32.1
regex==2023.12.25
requests==2.31.0
requests-gssapi==1.2.3
requests-kerberos==0.14.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.16.2
rsa==4.9
safetensors==0.4.1
scikit-image==0.21.0
scipy==1.11.2
seaborn==0.12.2
Send2Trash==1.8.2
setuptools-scm==8.0.4
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
sspilib==0.1.0
stack-data==0.6.2
sympy==1.12
tensorboard==2.14.0
tensorboard-data-server==0.7.1
terminado==0.18.0
thop==0.1.1.post2209072238
tifffile==2023.9.26
tinycss2==1.2.1
tokenizers==0.15.0
tomli==2.0.1
toolz==0.12.0
torch==2.0.0+cu117
torchaudio==2.0.1+cu117
torchvision==0.15.1+cu117
tornado==6.4
tqdm==4.66.1
traitlets==5.9.0
transformers==4.36.2
trimesh==3.23.5
types-python-dateutil==2.8.19.20240106
typing_extensions==4.7.1
tzdata==2023.3
ujson==5.9.0
uri-template==1.3.0
urllib3==1.26.16
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==3.0.1
whichcraft==0.6.1
widgetsnbextension==3.6.6
wrapt==1.16.0
y-py==0.6.2
ypy-websocket==0.8.4
zipp==3.16.2
zope.event==5.0
zope.interface==6.1
